requirements.txt
asgiref==3.7.2
Django==3.2.25
django-cors-headers==4.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
django-ratelimit==4.1.0
psycopg2-binary==2.9.9
python-dotenv==0.21.1
pytz==2025.2
sqlparse==0.4.4
typing_extensions==4.7.1
gunicorn==20.1.0
whitenoise==6.2.0
scikit-learn==1.3.0
numpy==1.24.3
scipy==1.10.1
joblib==1.3.2
drf-yasg==1.21.7
inflection==0.5.1
uritemplate==4.1.1
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
coverage==7.2.7
